Work has begun on the site.
html - added files "main.html", "projects.html" and "contacts.html". In each of these files, a common template was added for more convenient work. The <body> tags were given classes, for special sizes, for each page of the site.

css - in the file style.scss/css, was added import of the main font of the site, was added <body> tag, in it was built in the font and classes of the main .html files

assets - added basic images, to work with .html and style.scss/css files.

js - the files "greetings.js" and "status.js" were created.

status.js - responsible for the response state of the site from the server.

greetings.js - greetings when you enter the site.

@AidenOrtiz