Conducted intermediate work, in the "footer" block.
html - the footer block was added. Classes and tags were added to the block, and content was added to the site.

css - Classes and tags of the footer block, as well as the block itself were added to the file style.scss/css. Flex-elements, scss/css style properties were applied, positioning + adding new variable "$color5" was used for main.html block.

img - in the folder img/footer, content was added to further work with the bottom menu of the site.

@AidenOrtiz