Work on the "about-middle" block is finished. Functions were added to the previously created animations.
html - additional tags and classes were added to create a more beautiful design on the site. The class "about-middle__box" was added to create the borders of one of the lists.

scss/css - in the previously created animations, the animation attribute "transition" was applied for smooth animation, after using the pseudoclass "hover". Was finalized on the addition of the property styles, for the block "about-middle" and a new class of "about-middle__box"

img - in the folder img/main, was added an image.