Work has begun on the "projects-main.html" file. The file "projects.html" was replaced by "projects-main.html".
html - the main "projects" block has been added. Basic classes, tags and content were added to it. A class was created for the slider, which will be used later

scss/css - The "projects" block was added to the style.scss/css file.

img - images were added to the folder img/projects/slider, for future work with the slider.