Work on the "projects" block is finished.
html - The structure of the "project-slider" block was changed in the course of the work. New classes, blocks and tags were added. The original version of the "project-slider" block was completely removed.

scss/css - the structure of classes, tags and content was changed. Added new structure, defined basic page parameters, style properties, positioning and flex-elements + animations were added and slider was completely made.

img - in the folder img/projects/slider, new images were added in place of the old ones.