Intermediate work has been done in the "project-pages" block. Beginning work on the "project-pages-slider" block.
html - The "project-pages" and "project-pages-slider" were completely redone. Additional tags, classes and content were added. Additional id's were added for the future work of the scripts.

scss/css - most of the work with style properties, flex-elements and pseudo-elements was done in "project-pages". As the basis for the styles of "project-pages-slider" were taken by the previous developments.

img - in the folder img/projects/slider, images were added for further work.