Intermediate work on the "project-pages" block and on the "slider.js" file. The "sliderText.js" file was renamed "slider.js".
html - Unnecessary tags were commented out, redesigned, and new id's were added to classes and tags.

scss/css - The position of slider buttons was changed, because of removing unnecessary buttons.

js - additional variables were added to the js file. functions were added to the r1 and r2 events to replace tool icons and links when a slide is flipped.