Work on the "mobile-adaptive" branch (8).
html - The code structure was changed, obsolete links were removed and replaced by more relevant ones.

css/scss - In the course of the work, the "Projects-page.html" page was adapted. New style properties were used for the page, which will be used when working with mobile devices. All the content is now placed vertically

img - the image has been changed to a more up-to-date one

js - in the "slider.js" script, the obsolete links have been changed to the current ones.